+++ api/main.go
@@ -39,8 +39,8 @@ import (
 	"github.com/wormhole-foundation/wormhole-explorer/api/routes/wormscan"
 	rpcApi "github.com/wormhole-foundation/wormhole-explorer/api/rpc"
 	wormscanCache "github.com/wormhole-foundation/wormhole-explorer/common/client/cache"
+	"github.com/wormhole-foundation/wormhole-explorer/common/db"
 	xlogger "github.com/wormhole-foundation/wormhole-explorer/common/logger"
-	"github.com/wormhole-foundation/wormhole-explorer/common/mongohelpers"
 	"github.com/wormhole-foundation/wormhole-explorer/common/utils"
 	"go.uber.org/zap"
 )
@@ -103,7 +103,7 @@ func main() {
 
 	// Setup DB
 	rootLogger.Info("connecting to MongoDB")
-	db, err := mongohelpers.Connect(appCtx, cfg.DB.URL, cfg.DB.Name)
+	db, err := db.Connect(appCtx, cfg.DB.URL, cfg.DB.Name)
 	if err != nil {
 		rootLogger.Fatal("failed to connect to MongoDB", zap.Error(err))
 	}
